+++ src/io/segment/android/Configuration.java
@@ -10,6 +10,7 @@ public class Configuration {
 	private static final String INTEGER_RESOURCE_KEY = "integer";
 	
 	private static final String SECRET_KEY = "analytics_secret";
+	private static final String WRITE_KEY = "analytics_write_key";
 	
 	private static final String HOST_KEY = "analytics_host";
 	private static final String DEBUG_KEY = "analytics_debug";
@@ -19,8 +20,13 @@ public class Configuration {
 	private static final String SETTINGS_CACHE_EXPIRY_KEY = "analytics_settings_cache_expiry";
 
 	
-	public static String getSecret(Context context) {
-		return getString(context, SECRET_KEY);
+	public static String getWriteKey(Context context) {
+		String writeKey = getString(context, WRITE_KEY);
+		if (writeKey != null) { 
+			return writeKey;
+		} else {
+			return getString(context, SECRET_KEY);
+		} 
 	}
 	
 	public static Options getOptions(Context context) {